Do not link to ruby-doc.org
ruby-doc.org is a commercial site that serves ads, we have an official
Ruby documentation hosting site at docs.ruby-lang.org.

It's not well ranked in search engines but it's the only official host
of Ruby documentation and it can't be hijacked by random folks to serve
ads or potentially much more malicious content to unsuspecting folks
looking at the Try Ruby playground.