Fix loading code from URI
Previously, if you hit the "Copy URL" button and visit the URL in a
different tab, you would not get the code that you had in the editor.
For example, try https://try.ruby-lang.org/playground/#code=no&engine=cruby-3.2.0dev

Additionally, if you edit the code in the URL and hit enter, you'd get a
error in the JS console.

Both of these issue were caused by getting a string back that start with
a '#'. Delete this prefix.