Ignore warnings from `gem_prelude`
With the latest `wasm-wasi`, the following warnings are shown on TryRuby.

```
`RubyGems' were not loaded.
`error_highlight' was not loaded.
`did_you_mean' was not loaded.
`syntax_suggest' was not loaded.
```

This warnings come from `gem_prelude.rb`.
https://github.com/ruby/ruby/blob/master/gem_prelude.rb

I assume this this is because the current usage of `ruby.wasm`
on TryRuby don't consider RubyGems and fails to load it.

We should change the structure of how to use `ruby.wasm` for fixing
this. But, this is just warnings and there are no other issues.
And, it seems that this warnings would not be shown in Ruby 3.5
https://github.com/ruby/ruby/pull/13454/commits/4e303fd786bee1d92a41306d055f078a3c8c0cb8
https://github.com/ruby/ruby/pull/13454/commits/e9dadf133434b7282269b143ff2b20111707f4f1

So I fixed to just silence those warnings.